+++ lib/ui/home/appbar/drawer.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:studyapp/ui/about/about.dart';
 import 'package:studyapp/l10n/app_localizations.dart';
+import 'package:studyapp/ui/download/download_manager_page.dart';
 import 'package:studyapp/ui/search/search.dart';
 import 'package:studyapp/ui/settings/settings_page.dart';
 
@@ -50,6 +51,19 @@ class AppDrawer extends StatelessWidget {
                 Navigator.push(
                   context,
                   MaterialPageRoute(builder: (context) => const SearchPage()),
+                );
+              },
+            ),
+            ListTile(
+              title: Text(AppLocalizations.of(context)!.downloads),
+              leading: const Icon(Icons.download),
+              onTap: () {
+                Navigator.pop(context);
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const DownloadManagerPage(),
+                  ),
                 );
               },
             ),