+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.16.0
 
 - Swipe up/down shortcut on Book and Chapter buttons.
+- Add an audio download manager.
+- Stream audio by if no download exists.
 
 ## 0.15.0 - January 13, 2026